+++ SharePointFramework/ProjectExtensions/src/components/TemplateSelectDialog/useSelectionList.ts
@@ -9,7 +9,7 @@ import { useEffect, useState } from 'react'
  * @param onSelectionChanged On selection changed
  */
 export function useSelectionList(
-  selectedKeys: string[],
+  selectedKeys: (string | number)[],
   onSelectionChanged: (items: any[]) => void
 ) {
   const __selection = new Selection<any>({
@@ -22,7 +22,7 @@ export function useSelectionList(
 
   useEffect(() => {
     __selection.setChangeEvents(false)
-    selectedKeys.forEach((key) => __selection.setKeySelected(key, true, true))
+    selectedKeys.forEach((key) => __selection.setKeySelected(key as any, true, true))
     __selection.setChangeEvents(true)
     setSelection(__selection)
   }, [searchTerm])
